requirements.txt
PyYAML
cerberus
validators
rich